Use .human_attribute_name as a fallback for column names
It passes the option '_title_human_attribute_name' to the column as I
wasn't able to access the ActiveRecord class within the 'translate' method.

What do you think of this approach? I'm happy to write a test around it
if you think it's acceptable but I think there might be a better way to
do this :)

This is part of Issue #12